fix: Limiter cross-test influence
* Limiter instance started caching things,
  tests were influencing one another.
* Split tests into two modules, added
  fresh_limiter_instance module-level fixture
  to reset the limiter
